+++ packages/react-json-tree/jest.config.cjs
@@ -1,0 +1,12 @@
+module.exports = {
+  extensionsToTreatAsEsm: ['.ts', '.tsx'],
+  moduleNameMapper: {
+    '^(\\.{1,2}/.*)\\.js$': '$1',
+  },
+  transform: {
+    '^.+\\.tsx?$': [
+      'ts-jest',
+      { tsconfig: 'tsconfig.test.json', useESM: true },
+    ],
+  },
+};