+++ packages/redux-devtools-ui/.storybook/preview.tsx
@@ -18,7 +18,12 @@ const withThemeProvider = (Story, context) => (
 
 const preview: Preview = {
   parameters: {
-    actions: { argTypesRegex: '^on[A-Z].*' },
+    controls: {
+      matchers: {
+        color: /(background|color)$/i,
+        date: /Date$/i,
+      },
+    },
   },
   globalTypes: {
     theme: {