+++ packages/react-json-tree/CHANGELOG.md
@@ -1,5 +1,16 @@
 # Change Log
 
+## 0.19.0
+
+### Minor Changes
+
+- bbb1a40: Convert React packages to ESM
+
+### Patch Changes
+
+- Updated dependencies [bbb1a40]
+  - react-base16-styling@0.10.0
+
 ## 0.18.0
 
 ### Major Changes