+++ packages/redux-devtools-chart-monitor/CHANGELOG.md
@@ -1,5 +1,19 @@
 # Change Log
 
+## 5.0.2
+
+### Patch Changes
+
+- Updated dependencies [bbb1a40]
+  - react-base16-styling@0.10.0
+
+## 5.0.1
+
+### Patch Changes
+
+- Updated dependencies [191d419]
+  - d3-state-visualizer@3.0.0
+
 ## 5.0.0
 
 ### Major Changes